refactor(frontend): split playtest runtime helpers
